# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_discover_schema_projects__project_id__schema_discover_post": {
        "properties": {
          "files": {
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "required": [
          "files"
        ],
        "title": "Body_discover_schema_projects__project_id__schema_discover_post",
        "type": "object"
      },
...
      "DatasetOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "semantic_model": {
            "$ref": "#/components/schemas/SemanticModel"
          }
        },
        "required": [
          "id",
          "project_id",
          "name",
          "semantic_model",
          "created_at"
        ],
        "title": "DatasetOut",
        "type": "object"
      },
      "Entity": {
        "properties": {
          "fields": {
            "items": {
              "$ref": "#/components/schemas/SemanticField"
            },
            "title": "Fields",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "fields"
        ],
        "title": "Entity",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobOut": {
        "properties": {
          "dataset": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DatasetOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          }
        },
        "required": [
          "id",
          "project_id",
          "status"
        ],
        "title": "JobOut",
        "type": "object"
      },
      "JobStatus": {
        "enum": [
          "pending",
          "running",
          "done",
          "failed"
        ],
        "title": "JobStatus",
        "type": "string"
      },
...
      "Relationship": {
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "direction": {
            "title": "Direction",
            "type": "string"
          },
          "from_field": {
            "title": "From Field",
            "type": "string"
          },
          "to_field": {
            "title": "To Field",
            "type": "string"
          },
          "validated": {
            "default": true,
            "title": "Validated",
            "type": "boolean"
          }
        },
        "required": [
          "from_field",
          "to_field",
          "confidence",
          "direction"
        ],
        "title": "Relationship",
        "type": "object"
      },
...
      "SemanticField": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source_column": {
            "title": "Source Column",
            "type": "string"
          },
          "source_file": {
            "title": "Source File",
            "type": "string"
          },
          "status": {
            "default": "auto",
            "enum": [
              "auto",
              "needs_confirmation",
              "confirmed",
              "rejected"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "name",
          "source_column",
          "source_file",
          "confidence"
        ],
        "title": "SemanticField",
        "type": "object"
      },
      "SemanticModel": {
        "properties": {
          "entities": {
            "items": {
              "$ref": "#/components/schemas/Entity"
            },
            "title": "Entities",
            "type": "array"
          },
          "relationships": {
            "items": {
              "$ref": "#/components/schemas/Relationship"
            },
            "title": "Relationships",
            "type": "array"
          }
        },
        "required": [
          "entities",
          "relationships"
        ],
        "title": "SemanticModel",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Phase 5/6 -- FastAPI app wiring POC 1-4's pipelines behind HTTP routes, with auth + multi-tenancy (Postgres, JWT, RBAC, S3-compatible dataset storage).",
    "title": "InsightFlow Backend",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/projects/{project_id}/datasets": {
      "get": {
        "description": "A project's datasets, newest first -- so `[0]` is the one query/dashboard/chat actually run\nagainst (`dataset_deps.get_latest_dataset` resolves the same row).\n\nAdded in Phase 8 M3: without it the frontend could only ever see a semantic model in the\nresponse of the discovery job it just polled, so a page reload left the UI unable to tell\nwhether the project had any data at all. Returns full semantic models rather than a summary\nplus a detail route -- one upload is one row and a model is a few KB, so the simpler shape\ncovers both \"does this project have data\" and \"show me the current model\" in one request.",
        "operationId": "list_datasets_projects__project_id__datasets_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DatasetOut"
                  },
                  "title": "Response List Datasets Projects  Project Id  Datasets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Datasets",
        "tags": [
          "schema"
        ]
      }
    },
...
    "/projects/{project_id}/schema/discover": {
      "post": {
        "description": "Enqueues discovery instead of running it inline (Phase 7 M3) -- LLM-heavy relationship\nreasoning could take a while for a large/many-file upload, and doing that inside the HTTP\nrequest meant the client's connection had to stay open the whole time. Poll\nGET .../schema/jobs/{id} for the result; `run_schema_discovery()` itself, and every\nvalidation rule below, are unchanged from the synchronous version.",
        "operationId": "discover_schema_projects__project_id__schema_discover_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_discover_schema_projects__project_id__schema_discover_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discover Schema",
        "tags": [
          "schema"
        ]
      }
    },
    "/projects/{project_id}/schema/jobs/{job_id}": {
      "get": {
        "operationId": "get_discovery_job_projects__project_id__schema_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Discovery Job",
        "tags": [
          "schema"
        ]
      }
    }
  }
}